refactor: import the 2FA cancellation guard from its leaf module
